[SP-242] refactor: resolver 클래스명 및 패키지명 수정
