+++ src/main/java/com/cupid/jikting/common/config/AuthorizationConfig.java
@@ -1,6 +1,6 @@
 package com.cupid.jikting.common.config;
 
-import com.cupid.jikting.common.ui.AuthResolver;
+import com.cupid.jikting.common.resolver.AuthorizationResolver;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
@@ -12,10 +12,10 @@ import java.util.List;
 @Configuration
 public class AuthorizationConfig implements WebMvcConfigurer {
 
-    private final AuthResolver authResolver;
+    private final AuthorizationResolver authorizationResolver;
 
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(authResolver);
+        resolvers.add(authorizationResolver);
     }
 }
